requirements.txt

aiogram==2.25.1
aiohttp==3.8.6
python-dotenv
flask
requests
